test: isolate signed HVF lifecycle test
