+++ crates/hvf/tests/hvf_lifecycle.rs
@@ -1,0 +1,21 @@
+#[cfg(all(target_os = "macos", target_arch = "aarch64"))]
+#[test]
+fn creates_and_destroys_hvf_vcpu() {
+    use bangbang_hvf::HvfBackend;
+    use bangbang_runtime::VmBackend;
+
+    let mut backend = HvfBackend::new();
+
+    backend.create_vm().expect("VM should be created");
+    {
+        let vcpu = backend.create_vcpu().expect("vCPU should be created");
+        vcpu.destroy().expect("vCPU should be destroyed");
+    }
+    backend.destroy_vm().expect("VM should be destroyed");
+}
+
+#[cfg(not(all(target_os = "macos", target_arch = "aarch64")))]
+#[test]
+fn requires_macos_apple_silicon() {
+    panic!("signed HVF lifecycle tests require macOS Apple Silicon");
+}
